+++ src/main/java/basemod/patches/com/megacrit/cardcrawl/dungeons/AbstractDungeon/InitializeEventListPatch.java
@@ -1,12 +1,12 @@
 package basemod.patches.com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 
-import java.util.Map.Entry;
-
-import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
-import com.megacrit.cardcrawl.dungeons.*;
-import com.megacrit.cardcrawl.events.AbstractEvent;
-
 import basemod.BaseMod;
+import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
+import com.megacrit.cardcrawl.events.AbstractEvent;
+import javassist.CannotCompileException;
+import javassist.expr.ExprEditor;
+import javassist.expr.MethodCall;
 
 public class InitializeEventListPatch {
 	@SpirePatch(cls = "com.megacrit.cardcrawl.helpers.EventHelper", method = "getEvent")
@@ -17,23 +17,11 @@ public class InitializeEventListPatch {
 			
 			BaseMod.logger.info("Looking for custom events...");
 
-			//hahah this code is really fuckin bad
-			
 			try {
-				event = BaseMod.getEventList(BaseMod.EventPool.ANY).get(key);
-				if(event != null)
+				event = BaseMod.getEvent(key);
+				if (event != null) {
 					return event.newInstance();
-				event = BaseMod.getEventList(BaseMod.EventPool.THE_EXORDIUM).get(key);
-				if(event != null)
-					return event.newInstance();
-				event = BaseMod.getEventList(BaseMod.EventPool.THE_CITY).get(key);
-				if(event != null)
-					return event.newInstance();
-				event = BaseMod.getEventList(BaseMod.EventPool.THE_BEYOND).get(key);
-				if(event != null)
-					return event.newInstance();
-				
-				
+				}
 			} catch (InstantiationException | IllegalAccessException e) {
 				// TODO Auto-generated catch block
 				e.printStackTrace();
@@ -41,54 +29,50 @@ public class InitializeEventListPatch {
 			return __result;
 		}
 	}
-	
-	@SpirePatch(cls = "com.megacrit.cardcrawl.dungeons.Exordium", method = "initializeEventList")
-	public static class TheExordiumEventList{
-		@SuppressWarnings("static-access")
-		public static void Postfix(Exordium dungeon) {
-			
-			BaseMod.logger.info("BaseMod Adding Custom Events...");
-			for(Entry<String, Class<? extends AbstractEvent>> entry : BaseMod.getEventList(BaseMod.EventPool.THE_EXORDIUM).entrySet()) {
-				BaseMod.logger.info("BaseMod added " + entry.getKey() + " to the Exordium eventList.");
-				dungeon.eventList.add(entry.getKey());
+
+	@SpirePatch(
+			cls="com.megacrit.cardcrawl.dungeons.AbstractDungeon",
+			method=SpirePatch.CONSTRUCTOR,
+			paramtypes={
+					"java.lang.String",
+					"java.lang.String",
+					"com.megacrit.cardcrawl.characters.AbstractPlayer",
+					"java.util.ArrayList"
 			}
+	)
+	public static class EventList {
+		public static ExprEditor Instrument() {
+			return new ExprEditor() {
+				@Override
+				public void edit(MethodCall m) throws CannotCompileException {
+					if (m.getMethodName().equals("initializeEventList")) {
+						m.replace("{" +
+								"$_ = $proceed($$);" +
+								"basemod.patches.com.megacrit.cardcrawl.dungeons.AbstractDungeon.InitializeEventListPatch.EventList.Nested.Do();" +
+								"}");
+					}
+				}
+			};
 		}
-	}
-	
-	@SpirePatch(cls = "com.megacrit.cardcrawl.dungeons.TheCity", method = "initializeEventList")
-	public static class TheCityEventList{
-		@SuppressWarnings("static-access")
-		public static void Postfix(TheCity dungeon) {
-			
-			BaseMod.logger.info("BaseMod Adding Custom Events...");
-			for(Entry<String, Class<? extends AbstractEvent>> entry : BaseMod.getEventList(BaseMod.EventPool.THE_CITY).entrySet()) {
-				BaseMod.logger.info("BaseMod added " + entry.getKey() + " to the Exordium eventList.");
-				dungeon.eventList.add(entry.getKey());
-			}
-		}
-	}
-	
-	@SpirePatch(cls = "com.megacrit.cardcrawl.dungeons.TheBeyond", method = "initializeEventList")
-	public static class TheBeyondEventList{
-		@SuppressWarnings("static-access")
-		public static void Postfix(TheBeyond dungeon) {
-			
-			BaseMod.logger.info("BaseMod Adding Custom Events...");
-			for(Entry<String, Class<? extends AbstractEvent>> entry : BaseMod.getEventList(BaseMod.EventPool.THE_BEYOND).entrySet()) {
-				BaseMod.logger.info("BaseMod added " + entry.getKey() + " to the Exordium eventList.");
-				dungeon.eventList.add(entry.getKey());
+
+		public static class Nested {
+			public static void Do() {
+				BaseMod.logger.info("Adding Custom Events...");
+				for (String eventID : BaseMod.getEventList(AbstractDungeon.id).keySet()) {
+					BaseMod.logger.info("Added " + eventID + " to the " + AbstractDungeon.id + " eventList.");
+					AbstractDungeon.eventList.add(eventID);
+				}
 			}
 		}
 	}
 	
 	@SpirePatch(cls = "com.megacrit.cardcrawl.dungeons.AbstractDungeon", method = "initializeSpecialOneTimeEventList")
 	public static class SpecialEventList{
-		@SuppressWarnings("static-access")
 		public static void Postfix(AbstractDungeon dungeon) {
-			BaseMod.logger.info("BaseMod Adding Custom SpecialOneTimeEvents...");
-			for(Entry<String, Class<? extends AbstractEvent>> entry : BaseMod.getEventList(BaseMod.EventPool.ANY).entrySet()) {
-				BaseMod.logger.info("BaseMod added " + entry.getKey() + " to the specialOneTimeEventList.");
-				dungeon.eventList.add(entry.getKey());
+			BaseMod.logger.info("Adding Custom SpecialOneTimeEvents...");
+			for(String eventID : BaseMod.getEventList((String)null).keySet()) {
+				BaseMod.logger.info("Added " + eventID + " to the specialOneTimeEventList.");
+				AbstractDungeon.specialOneTimeEventList.add(eventID);
 			}
 		}
 	}